fix libs & prd build
