+++ lib/ring_buff/ring_buff.c
@@ -1,5 +1,7 @@
 #include "ring_buff.h"
 
+#if WSH_USE_COLLECTIONS_RING_BUFF
+
 #define RING_BUFF_DEF_TMO_MS 3
 
 /**
@@ -297,7 +299,7 @@ RET_STATE_t RingBuff_Buff_Search(RingBuff_t* pRingBuff, u32 timeout, u8* pMatchB
 
 	while (SYS_TICK_GET_MS_CNT() <= endSearchTime) {
 		if (pRingBuff->Count < buffLen) {
-			SYS_DELAY_MS(RING_BUFF_DEF_TMO_MS);	 //Wait more data
+			SYS_DELAY_MS(RING_BUFF_DEF_TMO_MS);  //Wait more data
 			continue;
 		} else {
 			while (*locTail != pMatchBuff[0]) {
@@ -492,7 +494,7 @@ RET_STATE_t RingBuff_Str_PeekBetween(RingBuff_t* pRingBuff, u8* pDstBuff, u16 ds
 	return retState;
 }
 
-//-----------------------------------------------------------------------------  ---------------------------------------------
+//------------------------------------------------------------------------------
 
 RET_STATE_t RingBuff_Ring2Line_CopyBeforeEntry(RingBuff_t* pRingBuff, u32 timeout, u8* pDstBuff,
 											   u16 dstBuffMaxSize, char* pStringEntry,
@@ -661,3 +663,5 @@ RET_STATE_t RingBuff_Str_SetPeekFirst(RingBuff_t* pRingBuff, u32 timeout, char**
 	*pMatchStr = -1;
 	return RET_STATE_ERR_TIMEOUT;
 }
+
+#endif /* WSH_USE_COLLECTIONS_RING_BUFF */
